% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettotalcoveragetransactions.R
\name{get.total.coverage.transactions}
\alias{get.total.coverage.transactions}
\title{CoverageR}
\usage{
get.total.coverage.transactions(locations, lon, lat, distance)
}
\arguments{
\item{locations}{Data Frame of locations}

\item{lon}{Longitude for point of interest eg: The starting point}

\item{lat}{Latitude for point of interest eg: The starting point}

\item{distance}{Value for the size of the radius in KM eg: 1.4 or 3.0}
}
\value{
Table containing coverage information.
}
\description{
Calculates whether locations are within a defined radius.
}
\examples{
\dontrun{

Example One

distances <- c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1)
walking_time <- function(i) {
 distance <- i
 print(i)
 output <- get.multi.coverage(hotel,stop_data,i)
 output <- output \%>\% cbind(stop_locations)
 output <- output \%>\% distinct
 output$distance <- distance
 return(output)
 }

 result <- map_df(distances, walking_time)


}

}
\author{
Chloe Morgan
}
